requirements.txt
djangorestframework==3.12.2
beautifulsoup4>=4.9
requests>=2.25